+++ src/setupProxy.js
@@ -7,9 +7,9 @@ module.exports = function(app){
     '/api',
     createProxyMiddleware({
       //如果后端服务在远程
-      target: 'http://8.134.73.52:5001/',
+      // target: 'http://8.134.73.52:5001/',
       //如果后端服务开在本地
-      // target: 'http://localhost:5001',
+      target: 'http://localhost:5001',
       changeOrigin: true,
     })
   )
