+++ src/views/PageContainer/SystemManage/MetaData/MetaData.css
@@ -1,47 +1,10 @@
-.card-container p {
-    margin: 0;
+/* tile uploaded pictures */
+.upload-list-inline .ant-upload-list-item {
+    float: left;
+    width: 200px;
+    margin-right: 8px;
   }
-  .card-container > .ant-tabs-card .ant-tabs-content {
-    height: 120px;
-    margin-top: -16px;
-  }
-  .card-container > .ant-tabs-card .ant-tabs-content > .ant-tabs-tabpane {
-    padding: 16px;
-    background: #fff;
-  }
-  .card-container > .ant-tabs-card > .ant-tabs-nav::before {
-    display: none;
-  }
-  .card-container > .ant-tabs-card .ant-tabs-tab,
-  [data-theme='compact'] .card-container > .ant-tabs-card .ant-tabs-tab {
-    background: transparent;
-    border-color: transparent;
-  }
-  .card-container > .ant-tabs-card .ant-tabs-tab-active,
-  [data-theme='compact'] .card-container > .ant-tabs-card .ant-tabs-tab-active {
-    background: #fff;
-    border-color: #fff;
-  }
-  #components-tabs-demo-card-top .code-box-demo {
-    padding: 24px;
-    overflow: hidden;
-    background: #f5f5f5;
-  }
-  [data-theme='compact'] .card-container > .ant-tabs-card .ant-tabs-content {
-    height: 120px;
-    margin-top: -8px;
-  }
-  [data-theme='dark'] .card-container > .ant-tabs-card .ant-tabs-tab {
-    background: transparent;
-    border-color: transparent;
-  }
-  [data-theme='dark'] #components-tabs-demo-card-top .code-box-demo {
-    background: #000;
-  }
-  [data-theme='dark'] .card-container > .ant-tabs-card .ant-tabs-content > .ant-tabs-tabpane {
-    background: #141414;
-  }
-  [data-theme='dark'] .card-container > .ant-tabs-card .ant-tabs-tab-active {
-    background: #141414;
-    border-color: #141414;
+  
+  .upload-list-inline [class*='-upload-list-rtl'] .ant-upload-list-item {
+    float: right;
   }
